+++ backend/data/src/main/kotlin/io/tolgee/formats/ImportMessageFormat.kt
@@ -1,6 +1,7 @@
 package io.tolgee.formats
 
 import io.tolgee.formats.android.`in`.JavaToIcuPlaceholderConvertor
+import io.tolgee.formats.importMessageFormatConvertor.MessageFormatImportDataConvertor
 import io.tolgee.formats.paramConvertors.`in`.RubyToIcuPlaceholderConvertor
 
 enum class ImportMessageFormat(
@@ -8,6 +9,7 @@ enum class ImportMessageFormat(
   val canContainIcu: Boolean = false,
   val placeholderConvertorFactory: (() -> ToIcuPlaceholderConvertor)?,
   val rootKeyIsLanguageTag: Boolean = false,
+  val importDataConvertorFactory: (() -> MessageFormatImportDataConvertor)? = null,
 ) {
   YAML_RUBY(
     pluralsViaNesting = true,
