+++ backend/data/src/main/kotlin/io/tolgee/formats/genericStructuredFile/in/StructuredRawDataConversionResult.kt
@@ -1,7 +1,6 @@
 package io.tolgee.formats.genericStructuredFile.`in`
 
 class StructuredRawDataConversionResult(
-  val key: String,
   val value: String?,
   val isPlural: Boolean?,
 )
