+++ backend/data/src/main/kotlin/io/tolgee/formats/genericStructuredFile/in/StructuredRawDataConvertor.kt
@@ -2,7 +2,6 @@ package io.tolgee.formats.genericStructuredFile.`in`
 
 interface StructuredRawDataConvertor {
   fun convert(
-    keyPrefix: String,
     rawData: Any?,
     projectIcuPlaceholdersEnabled: Boolean,
     convertPlaceholdersToIcu: Boolean,
