+++ backend/data/src/main/kotlin/io/tolgee/repository/dataImport/ImportFileRepository.kt
@@ -1,11 +1,23 @@
 package io.tolgee.repository.dataImport
 
-import io.tolgee.model.dataImport.Import
 import io.tolgee.model.dataImport.ImportFile
 import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.data.jpa.repository.Query
 import org.springframework.stereotype.Repository
 
 @Repository
 interface ImportFileRepository : JpaRepository<ImportFile, Long> {
-  fun deleteAllByImport(import: Import)
+  @Query(
+    """
+    select f from ImportFile f
+    where f.import.project.id = :projectId 
+    and f.import.author.id = :authorId
+    and f.id = :id
+  """,
+  )
+  fun finByProjectAuthorAndId(
+    projectId: Long,
+    authorId: Long,
+    id: Long,
+  ): ImportFile?
 }
